+++ src/Arrays/AppendedArrays.jl
@@ -70,7 +70,7 @@ end
 
 Base.IndexStyle(::Type{<:AppendedArray}) = IndexLinear()
 
-function Base.getindex(v::AppendedArray,i::Integer)
+@propagate_inbounds function Base.getindex(v::AppendedArray, i::Integer)
   l = length(v.a)
   if i > l
     v.b[i-l]
@@ -87,7 +87,7 @@ function array_cache(v::AppendedArray)
   (ca,cb)
 end
 
-function getindex!(cache,v::AppendedArray,i::Integer)
+@propagate_inbounds function getindex!(cache, v::AppendedArray, i::Integer)
   ca, cb = cache
   l = length(v.a)
   if i > l