+++ src/Arrays/Interface.jl
@@ -1,18 +1,22 @@
 """
-    getindex!(cache,a::AbstractArray,i...)
+    getindex!(cache, a::AbstractArray, i...)
 
 Returns the item of the array `a` associated with index `i`
-by (possibly) using the scratch data passed in the `cache` object.
+by (possibly) using the scratch data passed in the `cache` object. The cache is
+only used for single element access (`i::Union{Integer,CartesianIndex,end,begin}...`),
+otherwise `getindex` is used.
 
-It defaults to
+`getindex!` defaults to
 
-    getindex!(cache,a::AbstractArray,i...) = a[i...]
+    getindex!(cache, a::AbstractArray, i...) = a[i...]
 
 As for standard Julia arrays, the user needs to implement only one of the following signatures
 depending on the `IndexStyle` of the array.
 
     getindex!(cache,a::AbstractArray,i::Integer)
     getindex!(cache,a::AbstractArray{T,N},i::Vararg{Integer,N}) where {T,N}
+
+See also [`array_cache`](@ref), [`uses_hash!`](@ref), [`invalidate_cache!`](@ref).
 
 # Examples
 
@@ -39,16 +43,34 @@ end
 ```
 
 """
-getindex!(cache,a::AbstractArray,i...) = a[i...]
-getindex!(cache,a::AbstractArray,i::CartesianIndex) = getindex!(cache,a,Tuple(i)...)
-getindex!(cache,a::AbstractArray,i::Integer) = _getindex_1d!(IndexStyle(a),cache,a,i)
-function getindex!(cache,a::AbstractArray{T,N},i::Vararg{Integer,N}) where {T,N}
-  _getindex_nd!(IndexStyle(a),cache,a,CartesianIndex(i))
+@propagate_inbounds function getindex!(cache, a::AbstractArray, inds...)
+  _getindex!(IndexStyle(a), cache, a, inds...)
 end
-_getindex_1d!(s::IndexLinear,cache,a,i) = a[i]
-_getindex_1d!(s::IndexCartesian,cache,a,i) = _getindex_nd!(s,cache,a,CartesianIndices(a)[i])
-_getindex_nd!(s::IndexLinear,cache,a,i) = _getindex_1d!(s,cache,a,LinearIndices(a)[i])
-_getindex_nd!(s::IndexCartesian,cache,a,i) = a[i]
+
+# if the expected overload of getindex! has not been dispatched, fallback to uncached getindex
+@propagate_inbounds _getindex!(::IndexLinear, cache, a, i::Integer) = a[i]
+@propagate_inbounds function _getindex!(::IndexCartesian, cache,
+  a::AbstractArray{T,N}, inds::Vararg{Integer,N}) where {T,N}
+
+  a[inds...]
+end
+
+# otherwise, convert indices to a single element in `a` to the correct style and try again
+@propagate_inbounds function _getindex!(::IndexLinear, cache, a, inds::Union{Integer, CartesianIndex}...)
+  @boundscheck checkbounds(a, inds...)
+  @inbounds i = LinearIndices(a)[inds...]
+  @inbounds getindex!(cache, a, i)
+end
+
+@propagate_inbounds function _getindex!(::IndexCartesian, cache, a, inds::Union{Integer, CartesianIndex}...)
+  @boundscheck checkbounds(a, inds...)
+  @inbounds ci = CartesianIndices(a)[inds...]
+  @inbounds getindex!(cache, a, Tuple(ci)...)
+end
+
+# else, unsuported in place indexing
+@propagate_inbounds _getindex!(::IndexStyle, cache, a, inds...) = a[inds...]
+
 
 """
     invalidate_cache!(cache) -> nothing
@@ -283,7 +305,7 @@ function test_array(
   cache = array_cache(a)
   t = true
   for i in eachindex(a)
-    ai = getindex!(cache,a,i)
+    @inbounds ai = getindex!(cache,a,i)
     t = t && (typeof(ai) <: eltype(a))
     t = t && (typeof(ai) <: T)
   end